Keep a map of Unit::strings as Unit::strmap
Strings in the bytecode are referred to by their position in the
Unit::strings vector, profiling shows that this can have a significant
performance impact on parsing, because the vector is not sorted and
linear search is performed for every string addition to avoid creating
duplicates.

By keeping track of string locations in a hashmap (std::unordered_map),
this can be sped up significantly. I.e. the test case went from
> 20s to 0.04s. The price is to pay in memory for storing each string
in both data structures.