Update the CALL/RET semantics
This change makes the CALL/LCALL/CALLLBL store the next operation
address on the stack. RET/LRET in turn now return to the address
on stack, instead of the operation following the operation at the
given address.

This change allows RET/LRET to return to 0. With the previous semantics
a return to 0 would always skip the first operation.